Review Comments for Initial Commit
The initial review of this PR concentrated on two areas that have been updated:
1) Whether to use require_relative or require, as well as not needing to use
   unnecessary requires.
2) File naming conventions for classes spanning multiple source files, now
   using a mixin file to merge the FS class from multiple files with different
   modules.
