+++ lib/virtfs/xfs/allocation_group.rb
@@ -2,7 +2,6 @@ require 'binary_struct'
 require 'virt_disk/disk_uuid'
 require 'stringio'
 require 'memory_buffer'
-require_relative 'superblock'
 
 require 'rufus/lru'
