+++ lib/virtfs/xfs/inode_map.rb
@@ -1,5 +1,3 @@
-require_relative 'inode'
-
 module Virtfs::XFS
   # ////////////////////////////////////////////////////////////////////////////
   # // Class.
