+++ lib/virtfs/xfs/short_form_directory_entry.rb
@@ -1,5 +1,4 @@
 require 'binary_struct'
-require_relative 'inode'
 
 module Virtfs::XFS
   # ////////////////////////////////////////////////////////////////////////////
